% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_passive}
\alias{dfr_passive}
\title{Passive dummy detector}
\usage{
dfr_passive()
}
\value{
Drifter object
}
\description{
Implements a dummy detector that always reports a drift. Useful
as an upper baseline, since it forces the model to be retrained at every
batch.
}
\examples{
model <- dfr_passive()
output <- update_state(model, 1)
output$drift
}
