% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dummy}
\alias{dummy}
\title{Dummy Drifter sub-class}
\usage{
dummy()
}
\value{
Drifter object
}
\description{
Ancestor class for the baseline detectors that ignore the
monitored signal, \code{dfr_inactive()} and \code{dfr_passive()}.
}
\examples{
obj <- dummy()
class(obj)
}
