% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{error_based}
\alias{error_based}
\title{Error Based Drifter sub-class}
\usage{
error_based()
}
\value{
Drifter object
}
\description{
Implements Error Based drift detectors. These detectors monitor
the residuals of a predictive model, where \code{1} (or \code{TRUE}) means a wrong
prediction and \code{0} (or \code{FALSE}) a correct one. Missing values are treated as
correct predictions.
}
\examples{
obj <- error_based()
class(obj)
}
