% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_manager.R
\name{fit.manager}
\alias{fit.manager}
\title{Process Batch}
\usage{
\method{fit}{manager}(obj, x, y, ...)
}
\arguments{
\item{obj}{manager object}

\item{...}{opitional arguments}

\item{data}{data batch in data frame format}

\item{prediction}{prediction batch as vector format}
}
\value{
updated Manager object
}
\description{
Process Batch
}
