% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_manager.R
\name{manager}
\alias{manager}
\title{Manager}
\usage{
manager(
  algoritm,
  n_slots = 1,
  model_selector = model_selector,
  model_deleter = model_deleter
)
}
\value{
Manager object
}
\description{
Ancestor class for Model Management
}
\examples{
# See ?dd_ddm for an example of DDM drift detector
}
