% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_accuracy}
\alias{mt_accuracy}
\title{Accuracy Calculator}
\usage{
mt_accuracy()
}
\value{
Metric object
}
\description{
Class for accuracy calculation
}
\examples{
library(daltoolbox)
obj <- mt_accuracy()
evaluate(obj, c(TRUE, FALSE, TRUE), c(TRUE, TRUE, TRUE))
}
