% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_fscore}
\alias{mt_fscore}
\title{FScore Calculator}
\usage{
mt_fscore(f = 1)
}
\arguments{
<<<<<<< HEAD
\item{f}{The beta parameter of the F-beta score. \code{f = 1} (the default) gives the usual F1 score, values below 1 weight precision more heavily and values above 1 weight recall more heavily.}
=======
\item{f}{The beta parameter for the F-Score metric}
>>>>>>> 264a2e411549c065608d2ace35dec5ebfba3725e
}
\value{
Metric object
}
\description{
Class for F-Score calculation. The \code{f} parameter is the beta of
the F-beta score, so \code{f = 1} (the default) gives the usual F1 score, values
below 1 weight precision more heavily and values above 1 weight recall more
heavily.
}
\examples{
library(daltoolbox)
obj <- mt_fscore(f = 1)
evaluate(obj, c(TRUE, FALSE, TRUE), c(TRUE, TRUE, TRUE))
}
