% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_recall}
\alias{mt_recall}
\title{Recall Calculator}
\usage{
mt_recall()
}
\value{
Metric object
}
\description{
Class for recall calculation
}
\examples{
library(daltoolbox)
obj <- mt_recall()
evaluate(obj, c(TRUE, FALSE, TRUE), c(TRUE, TRUE, TRUE))
}
