% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_rocauc}
\alias{mt_rocauc}
\title{ROC AUC Calculator}
\usage{
mt_rocauc()
}
\value{
Metric object
}
\description{
Class for ROC AUC calculation
}
\examples{
library(daltoolbox)
obj <- mt_rocauc()
evaluate(obj, c(0.9, 0.2, 0.8, 0.1), factor(c(TRUE, TRUE, FALSE, FALSE)))
}
