% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_norm.R
\name{norm}
\alias{norm}
\title{Norm}
\usage{
norm(norm_class)
}
\arguments{
\item{norm_class}{Normalizer class}
}
\value{
Norm object
}
\description{
Ancestor class for normalization techniques
}
\examples{
# See ?norm for an example of DDM drift detector
}
