% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_norm.R
\name{nrm_base}
\alias{nrm_base}
\title{Normalizer base class}
\usage{
nrm_base(norm_class)
}
\arguments{
\item{norm_class}{Normalizer class}
}
\value{
Norm object
}
\description{
Ancestor class for normalization techniques.
}
\details{
This constructor was named \code{norm()} up to version 1.2.x. It was
renamed to \code{nrm_base()} because \code{norm()} masks \code{\link[base:norm]{base::norm()}} once the
package is attached.
}
\examples{
library(daltoolbox)
obj <- nrm_base(norm_class = minmax())
class(obj)
}
