% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_manager.R
\name{predict.manager}
\alias{predict.manager}
\title{Predict Batch}
\usage{
\method{predict}{manager}(object, data, ...)
}
\arguments{
\item{data}{data batch in data frame format}

\item{...}{opitional arguments}

\item{obj}{manager object}

\item{prediction}{prediction batch as vector format}
}
\value{
predicted data
}
\description{
Predict Batch
}
