% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{reset_state}
\alias{reset_state}
\title{Reset State}
\usage{
reset_state(obj)
}
\arguments{
\item{obj}{Drifter object}
}
\value{
updated Drifter object
}
\description{
Reset Drifter State
}
\examples{
model <- dfr_ddm()
model <- reset_state(model)
model$drifted
}
