% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_selector.R
\name{selector}
\alias{selector}
\title{Selector}
\usage{
selector(slots_list)
}
\value{
Selector object
}
\description{
Ancestor class for model selection
}
\examples{
# See ?dd_ddm for an example of DDM drift detector
}
