% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_real_examples}
\alias{st_real_examples}
\title{Real time series for concept drift detection}
\format{
A list of real multivariate time series.
}
\source{
\href{https://github.com/cefet-rj-dal/heimdall}{Stealthy package}
}
\usage{
data(st_real_examples)
}
\description{
A list of real multivariate time series for concept drift detection
\itemize{
\item bfd1: Brazilian Flights Data 2023
}
#'
}
\examples{
data(st_real_examples)
dataset <- st_real_examples$bfd1
}
\references{
\href{https://github.com/cefet-rj-dal/heimdall}{Stealthy package}
}
\keyword{datasets}
