% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{update_state}
\alias{update_state}
\title{Update State}
\usage{
update_state(obj, value, ...)
}
\arguments{
\item{obj}{Drifter object}

\item{value}{a value that represents a processed batch}

\item{...}{optional arguments}
}
\value{
a list with two elements: \code{obj}, the updated Drifter object, and
\code{drift}, a logical flag indicating whether a drift was detected in this call
}
\description{
Update Drifter State with a new observation.
}
\examples{
model <- dfr_ddm()
output <- update_state(model, 0)
output$drift
}
