% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_adwin.R
\name{dfr_adwin}
\alias{dfr_adwin}
\title{ADWIN method}
\usage{
dfr_adwin(target_feat, delta = 0.002)
}
\arguments{
\item{target_feat}{Feature to be monitored.}

\item{delta}{The significance parameter for the ADWIN algorithm.}
}
\value{
\code{dfr_adwin} object
}
\description{
Adaptive Windowing method for concept drift detection \url{doi:10.1137/1.9781611972771.42}.
}
\examples{
library(daltoolbox)
library(heimdall)

# This example assumes a model residual where 1 is an error and 0 is a correct prediction.

data(st_drift_examples)
data <- st_drift_examples$univariate
data$event <- NULL
data$prediction <- st_drift_examples$univariate$serie > 4


model <- dfr_adwin(target_feat='serie')

detection <- c()
output <- list(obj=model, pred=FALSE)
for (i in 1:length(data$serie)){
 output <- update_state(output$obj, data$serie[i])
 if (output$pred){
   type <- 'drift'
   output$obj <- reset_state(output$obj)
 }else{
   type <- ''
 }
 detection <- rbind(detection, list(idx=i, event=output$pred, type=type))
}

detection <- as.data.frame(detection)
detection[detection$type == 'drift',]
}
