% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_aedd.R
\name{dfr_aedd}
\alias{dfr_aedd}
\title{Autoencoder-Based Drift Detection method}
\usage{
dfr_aedd(
  features,
  input_size,
  encoding_size,
  batch_size = 32,
  num_epochs = 1000,
  learning_rate = 0.001,
  window_size = 100,
  monitoring_step = 1700,
  criteria = "mann_whitney"
)
}
\arguments{
\item{features}{Features to be monitored}

\item{input_size}{Input size}

\item{encoding_size}{Encoding Size}

\item{batch_size}{Batch Size for batch learning}

\item{num_epochs}{Number of Epochs for training}

\item{learning_rate}{Learning Rate}

\item{window_size}{Size of the most recent data to be used}

\item{monitoring_step}{The number of rows that the drifter waits to be is updated}
}
\value{
\code{dfr_aedd} object
}
\description{
Autoencoder-Based method for concept drift detection \url{doi:0.1109/ICDMW58026.2022.00109}.
}
