% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_inactive}
\alias{dfr_inactive}
\title{Inactive dummy detector}
\usage{
dfr_inactive()
}
\value{
Drifter object
}
\description{
Implements a dummy detector that never reports a drift. Useful
<<<<<<< HEAD
as a lower baseline when evaluating adaptation strategies.
=======
as a baseline when evaluating adaptation strategies.
>>>>>>> 264a2e411549c065608d2ace35dec5ebfba3725e
}
\examples{
model <- dfr_inactive()
output <- update_state(model, 1)
output$drift
}
