% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_inactive}
\alias{dfr_inactive}
\title{Inactive dummy detector}
\usage{
dfr_inactive()
}
\value{
Drifter object
}
\description{
Implements a dummy detector that never reports a drift. Useful
as a baseline when evaluating adaptation strategies.
}
\examples{
model <- dfr_inactive()
output <- update_state(model, 1)
output$drift
}
