% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_inactive}
\alias{dfr_inactive}
\title{Inactive dummy detector}
\usage{
dfr_inactive()
}
\value{
Drifter object
}
\description{
Implements Inactive Dummy Detector
}
\examples{
# See ?hcd_ddm for an example of DDM drift detector
}
