% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_lbdd.R
\name{dfr_lbdd}
\alias{dfr_lbdd}
\title{Levene Based Drift Detection Method method}
\usage{
dfr_lbdd(target_feat = NULL, alpha = 0.01, window_size = 1500)
}
\arguments{
\item{target_feat}{Feature to be monitored}

\item{alpha}{Probability theshold for the test statistic}

\item{window_size}{Size of the sliding window}
}
\value{
\code{dfr_lbdd} object
}
\description{
Variation comparison using Levene Test for concept drift detection.
}
\examples{
library(daltoolbox)
library(heimdall)

# This example uses a dist-based drift detector with a synthetic dataset.

data(st_drift_examples)
data <- st_drift_examples$univariate
data$event <- NULL

model <- dfr_lbdd(target_feat='depart_visibility')

detection <- NULL
output <- list(obj=model, drift=FALSE)
for (i in 1:length(data$serie)){
 output <- update_state(output$obj, data$serie[i])
 if (output$drift){
   type <- 'drift'
   output$obj <- reset_state(output$obj)
 }else{
   type <- ''
 }
 detection <- rbind(detection, data.frame(idx=i, event=output$drift, type=type))
}

detection[detection$type == 'drift',]
}
