% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_mcdd.R
\name{dfr_mcdd}
\alias{dfr_mcdd}
\title{Mean Comparison Distance method}
\usage{
dfr_mcdd(target_feat = NULL, alpha = 1e-08, window_size = 1500)
}
\arguments{
\item{target_feat}{Feature to be monitored}

\item{alpha}{Probability theshold for all test statistics}

\item{window_size}{Size of the sliding window}
}
\value{
\code{dfr_mcdd} object
}
\description{
Mean Comparison statistical method for concept drift detection.
}
\examples{
library(daltoolbox)
library(heimdall)

# This example uses a dist-based drift detector with a synthetic dataset.

data(st_drift_examples)
data <- st_drift_examples$univariate
data$event <- NULL

model <- dfr_mcdd(target_feat='depart_visibility')

detection <- NULL
output <- list(obj=model, drift=FALSE)
for (i in 1:length(data$serie)){
 output <- update_state(output$obj, data$serie[i])
 if (output$drift){
   type <- 'drift'
   output$obj <- reset_state(output$obj)
 }else{
   type <- ''
 }
 detection <- rbind(detection, data.frame(idx=i, event=output$drift, type=type))
}

detection[detection$type == 'drift',]
}
