% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_multi_criteria}
\alias{dfr_multi_criteria}
\title{Multi Criteria Drifter sub-class}
\usage{
dfr_multi_criteria(drifter_list, combination = "or", fuzzy_window = 10)
}
\value{
Drifter object
}
\description{
Implements Multi Criteria drift detectors
}
