% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dfr_passive}
\alias{dfr_passive}
\title{Passive dummy detector}
\usage{
dfr_passive()
}
\value{
Drifter object
}
\description{
Implements Passive Dummy Detector
}
\examples{
# See ?hcd_ddm for an example of DDM drift detector
}
