% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dist_based}
\alias{dist_based}
\title{Distribution Based Drifter sub-class}
\usage{
dist_based(target_feat)
}
\arguments{
\item{target_feat}{Feature to be monitored.}
}
\value{
Drifter object
}
\description{
Implements univariate Distribution Based drift detectors. These
detectors monitor a single numeric feature; observations that are missing are
skipped instead of being imputed.
}
\examples{
obj <- dist_based(target_feat = 'serie')
obj$target_feat
}
