% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{drifter}
\alias{drifter}
\title{Drifter}
\usage{
drifter(detec_cooldown = 0)
}
\value{
Drifter object
}
\description{
Ancestor class for drift detection
}
\examples{
# See ?dd_ddm for an example of DDM drift detector
}
