% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{dummy}
\alias{dummy}
\title{Dummy Drifter sub-class}
\usage{
dummy()
}
\value{
Drifter object
}
\description{
Implements Error Based drift detectors
}
\examples{
# See ?hcd_ddm for an example of DDM drift detector
}
