% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{fit.drifter}
\alias{fit.drifter}
\title{Process Batch}
\usage{
\method{fit}{drifter}(obj, data, prediction, ...)
}
\arguments{
\item{obj}{Drifter object}

\item{data}{data batch in data frame format}

\item{prediction}{prediction batch as vector format}

\item{...}{opitional arguments}
}
\value{
updated Drifter object
}
\description{
Process Batch
}
