% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_accuracy}
\alias{mt_accuracy}
\title{Accuracy Calculator}
\usage{
mt_accuracy()
}
\value{
Metric object
}
\description{
Class for accuracy calculation
}
\examples{
# See ?mt_accuracy for an example of Accuracy Calculator
}
