% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_fscore}
\alias{mt_fscore}
\title{FScore Calculator}
\usage{
mt_fscore(f = 1)
}
\arguments{
\item{f}{The beta parameter of the F-beta score. \code{f = 1} (the default) gives the usual F1 score, values below 1 weight precision more heavily and values above 1 weight recall more heavily.}
}
\value{
Metric object
}
\description{
Class for FScore calculation
}
\examples{
# See ?mt_fscore for an example of FScore Calculator
}
