% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_fscore}
\alias{mt_fscore}
\title{FScore Calculator}
\usage{
mt_fscore(f = 1)
}
\arguments{
\item{f}{The F parameter for the F-Score metric}
}
\value{
Metric object
}
\description{
Class for FScore calculation
}
\examples{
# See ?mt_fscore for an example of FScore Calculator
}
