% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_precision}
\alias{mt_precision}
\title{Precision Calculator}
\usage{
mt_precision()
}
\value{
Metric object
}
\description{
Class for precision calculation
}
\examples{
# See ?mt_precision for an example of Precision Calculator
}
