% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_metrics.R
\name{mt_rocauc}
\alias{mt_rocauc}
\title{ROC AUC Calculator}
\usage{
mt_rocauc()
}
\value{
Metric object
}
\description{
Class for QOC AUC calculation
}
\examples{
# See ?mt_rocauc for an example of ROC AUC Calculator
}
