% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{mv_dist_based}
\alias{mv_dist_based}
\title{Multivariate Distribution Based Drifter sub-class}
\usage{
mv_dist_based()
}
\value{
Drifter object
}
\description{
Implements Multivariate Distribution Based drift detectors
}
