% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_norm.R
\name{nrm_memory}
\alias{nrm_memory}
\title{Memory Normalizer}
\usage{
nrm_memory(norm_class = minmax())
}
\value{
Norm object
}
\description{
Normalizer that has own memory
}
\examples{
# See ?nrm_mimax for an example of Memory Normalizer
}
