% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_drift_examples}
\alias{st_drift_examples}
\title{Synthetic time series for concept drift detection}
\format{
A list of data frames.
}
\source{
\href{https://github.com/cefet-rj-dal/heimdall}{heimdall package}
}
\usage{
data(st_drift_examples)
}
\description{
<<<<<<< HEAD
A list of univariate and multivariate time series for drift
detection. Each element is a data frame:
\itemize{
\item \code{uv_virtual_drift}: univariate series with a change in the feature distribution
\item \code{uv_vct_real_drift}: univariate series with a virtual change followed by a real drift
\item \code{uv_vdp_real_drift}: univariate series with a real drift and no change in the feature distribution
\item \code{mv_real_drift}: bivariate series with a real drift
\item \code{mv_vct_real_drift}: bivariate series with a virtual change followed by a real drift
=======
A list of time series for drift detection. It contains, among
others:
\itemize{
\item \code{example1}: a bivariate dataset with one multivariate concept drift example
\item \code{univariate}: a univariate series with four known change points
>>>>>>> 264a2e411549c065608d2ace35dec5ebfba3725e
}
}
\examples{
data(st_drift_examples)
dataset <- st_drift_examples$uv_virtual_drift
}
\references{
\href{https://github.com/cefet-rj-dal/heimdall}{heimdall package}
}
\keyword{datasets}
