% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_drift_examples}
\alias{st_drift_examples}
\title{Synthetic time series for concept drift detection}
\format{
A list of time series.
}
\source{
\href{https://github.com/cefet-rj-dal/heimdall}{heimdall package}
}
\usage{
data(st_drift_examples)
}
\description{
A list of time series for drift detection. It contains, among
others:
\itemize{
\item \code{example1}: a bivariate dataset with one multivariate concept drift example
\item \code{univariate}: a univariate series with four known change points
}
}
\examples{
data(st_drift_examples)
dataset <- st_drift_examples$example1
}
\references{
\href{https://github.com/cefet-rj-dal/heimdall}{heimdall package}
}
\keyword{datasets}
