% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_drift_examples}
\alias{st_drift_examples}
\title{Synthetic time series for concept drift detection}
\format{
A list of time series.
}
\source{
\href{https://github.com/cefet-rj-dal/heimdall}{Stealthy package}
}
\usage{
data(st_drift_examples)
}
\description{
A list of multivariate time series for drift detection
\itemize{
\item example1: a bivariate dataset with one multivariate concept drift example
}
#'
}
\examples{
data(st_drift_examples)
dataset <- st_drift_examples$uv_virtual_drift
}
\references{
\href{https://github.com/cefet-rj-dal/heimdall}{Stealthy package}
}
\keyword{datasets}
