% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_stealthy.R
\name{stealthy}
\alias{stealthy}
\title{Stealthy}
\usage{
stealthy(
  model,
  drift_method,
  th = 0.5,
  target_uni_drifter = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{model}{The algorithm object to be used for predictions}

\item{drift_method}{The algorithm object to detect drifts}

\item{th}{The threshold to be used with classification algorithms}

\item{verbose}{if TRUE shows drift messages}
}
\value{
Stealthy object
}
\description{
Ancestor class for drift adaptive models
}
\examples{
# See ?dd_ddm for an example of DDM drift detector
}
