% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_drifter.R
\name{update_state}
\alias{update_state}
\title{Update State}
\usage{
update_state(obj, value)
}
\arguments{
\item{obj}{Drifter object}

\item{value}{a value that represents a processed batch}
}
\value{
updated Drifter object
}
\description{
Update Drifter State
}
\examples{
# See ?hcd_ddm for an example of DDM drift detector
}
